\name{panel.qq}
\alias{panel.qq}
\title{ Panel function for qq}
\description{
  This is the default panel function for qq
}
\usage{
panel.qq(x, y, f.value=ppoints, ...)
}
\arguments{
  \item{x}{ Numeric variable}
  \item{y}{ Numeric variable with 2 unique values }
  \item{f.value}{ function for producing quantiles }
  \item{\dots}{ other arguments }
}
\author{Deepayan Sarkar \email{deepayan@stat.wisc.edu}}
\seealso{\code{\link{qq}}}
\keyword{dplot }
