\name{panel.tmd}
\alias{panel.tmd}
\title{ Panel function for tmd}
\description{
  This is the default panel function for tmd. Essentially panel.xyplot,
  with a horizontal line at y = 0.
}
\usage{
panel.tmd(...)
}
\arguments{
  \item{\dots}{ arguments passed to \code{panel.xyplot}}
}
\author{Deepayan Sarkar \email{Deepayan.Sarkar@R-project.org}}
\seealso{\code{\link{tmd}} ,\code{\link{Lattice}}}
\keyword{dplot }

