\name{trellis.datasets}
\alias{barley}
\alias{singer}
\alias{melanoma}
\alias{ethanol}
\alias{sunspot}
\alias{environmental}
\non_function{}
\title{Data Sets in the Lattice library}
\usage{
data(singer)
data(melanoma)
data(ethanol)
data(sunspot)
data(environmental)
}
\description{
  These data sets are included to facilitate the same examples as S-PLUS
  trellis. (Detailed descriptions of the datasets are omitted for now.)
}
\source{
  These datasets are part of the collection of datasets used in the book
  \emph{Visualizing Data} by Bill Cleveland, and are available at Bill
  Cleveland's Homepage
  \url{http://cm.bell-labs.com/cm/ms/departments/sia/wsc/index.html}
  as well as in Statlib.
}
\keyword{datasets}







