\name{panel.densityplot}
\alias{panel.densityplot}
\title{ Default Panel Function for densityplot }
\description{
  This is the default panel function for \code{densityplot}.
}
\usage{
panel.densityplot(x, n = 50, ...)
}
\arguments{
  \item{x}{ data points for which density is to be estimated}
  \item{n}{ number of (equally spaced) points to use for plotting}
  \item{\dots}{ parameters to be passed to the function \code{density},
    as well as graphical parameters like \code{col}}
}
\author{ Deepayan Sarkar \email{deepayan@stat.wisc.edu}}
\seealso{\code{\link{densityplot}}}
\keyword{dplot}
















