\name{panel.dotplot}
\alias{panel.dotplot}
\title{ Default Panel Functions for dotplot}
\description{
  This is the default panel function for \code{dotplot}.
}
\synopsis{
panel.dotplot(x, y, pch, col, ...)
}
\usage{
panel.dotplot(x, y, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x,y}{ coordinates of points to plot}
  \item{\dots}{ graphical parameters can be supplied}
}
\details{ Creates Dotplot of \code{x} against \code{y}
}
\seealso{\code{\link{dotplot}}}
\author{ Deepayan Sarkar \email{deepayan@stat.wisc.edu}}
\keyword{dplot}
















