\name{trellis.datasets}
\alias{singer}
\alias{melanoma}
\alias{ethanol}
\alias{environmental}
\docType{data}
\title{Data Sets in the Lattice library}
\usage{
singer
melanoma
ethanol
environmental
}

\description{
  These are data sets included in the \code{lattice} package that are
  not yet documented separately.  See the 'See Also' section for the
  ones that are. 
}

\source{
  These datasets are part of the collection of datasets used in the book
  \emph{Visualizing Data} by Bill Cleveland, and are available at 
  \url{http://cm.bell-labs.com/cm/ms/departments/sia/wsc/index.html}
  as well as in Statlib.
}
\seealso{
  \code{\link{barley}}
}
\keyword{datasets}







